pyproject.toml
[tool.poetry]
name = "listenbrainz_to_sqlite"
version = "0.1.0"
description = ""
authors = ["heyarne <arne@schlueter.is>"]
license = "GPL"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
click = "^8.0.3"
tqdm = "^4.62.3"
arrow = "^1.2.0"
yoyo-migrations = "^7.3.2"

[tool.poetry.dev-dependencies]
bpython = "^0.21"
datasette = "^0.59"
datasette-vega = "^0.6.2"
datasette-saved-queries = "^0.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
import_listens = "listenbrainz_to_sqlite:import_listens"
